Changed DataIds to Be 16-bit

+++ RED/Contexts/CommandMetadataContext.cs
@@ -9,7 +9,7 @@ namespace RED.Contexts
     [Serializable]
     public class CommandMetadataContext : IMetadata
     {
-        public byte Id { get; set; }
+        public ushort Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public string Datatype { get; set; }
@@ -31,7 +31,7 @@ namespace RED.Contexts
         {
             return new JsonCommandMetadataContext()
             {
-                Cmd_ID = Id,
+                Cmd_ID = (byte)Id,
                 Name = Name,
                 Datatype = Datatype,
                 Description = Description
