+++ RED/Contexts/TelemetryMetadataContext.cs
@@ -9,7 +9,7 @@ namespace RED.Contexts
     [Serializable]
     public class TelemetryMetadataContext : IMetadata
     {
-        public byte Id { get; set; }
+        public ushort Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public string Datatype { get; set; }
@@ -37,7 +37,7 @@ namespace RED.Contexts
         {
             return new JsonTelemetryMetadataContext()
             {
-                Telem_ID = Id,
+                Telem_ID = (byte)Id,
                 Name = Name,
                 Description = Description,
                 Datatype = Datatype,
